requirements.txt
altgraph==0.17
future==0.18.2
pefile==2021.5.13
pycryptodome==3.10.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pywin32==300
pywin32-ctypes==0.2.0
